Add 'follow topic' feature Fixes #55
